swaraj working

+++ Meeting-Management-System/src/main/java/Repository/GenderRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GenderRepo extends JpaRepository<Gender, Integer> {
+public interface GenderRepo extends JpaRepository<Gender, Short> {
 }
