+++ Meeting-Management-System/src/main/java/Repository/MeetAttendanceRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MeetAttendanceRepo extends JpaRepository<MeetAttendance, Integer> {
+public interface MeetAttendanceRepo extends JpaRepository<MeetAttendance, Long> {
 }
