+++ Meeting-Management-System/src/main/java/Repository/MinutTypeRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MinutTypeRepo extends JpaRepository<MinutType, Integer> {
+public interface MinutTypeRepo extends JpaRepository<MinutType, Long> {
 }
